requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pdf2image==1.16.3
Pillow==10.0.0
